replace non-base ASCII operators by their (very) verbose version
This is because Apple's Script Editor is just a crazy piece of crap
and replaces some operators by their non-base ASCII version.
This makes the code very hard to read on a regular text editor.
Thus, I've replaced these operators by their verbose version.
For example, the "<=" (in one byte) has been replaced by the text
"is less than or equal to".

For more information check the AppleScript's Operators Reference:
https://developer.apple.com/library/archive/documentation/AppleScript/Conceptual/AppleScriptLangGuide/reference/ASLR_operators.html